Refactor: formatting, files rename, common code abstraction and debug log
