+++ cpp/arcticdb/storage/s3/s3_mock_client.hpp
