+++ cpp/arcticdb/storage/test/test_azure_storage.cpp
@@ -12,7 +12,7 @@
 #include <arcticdb/storage/azure/azure_storage.hpp>
 #include <arcticdb/storage/azure/azure_mock_client.hpp>
 #include <arcticdb/util/buffer.hpp>
-#include <arcticdb/stream/test/stream_test_common.hpp>
+#include <arcticdb/storage/test/common.hpp>
 
 using namespace arcticdb;
 using namespace storage;
@@ -30,52 +30,6 @@ protected:
     AzureStorage store;
     AzureMockStorageFixture() : store(LibraryPath("_arctic_cfg", '.'), OpenMode::DELETE, get_mock_azure_config()) {}
 };
-
-VariantKey get_variant_key(std::string name) {
-    auto builder = atom_key_builder();
-    return builder.build<KeyType::TABLE_DATA>(name);
-}
-
-Segment get_example_segment(){
-    auto segment_in_memory = get_test_frame<arcticdb::stream::TimeseriesIndex>("symbol", {}, 10, 0).segment_;
-    auto codec_opts = proto::encoding::VariantCodec();
-    return encode_dispatch(std::move(segment_in_memory), codec_opts, EncodingVersion::V2);
-}
-
-void write_in_store(AzureStorage& store, std::string symbol) {
-    auto variant_key = get_variant_key(symbol);
-    store.write(KeySegmentPair(std::move(variant_key), get_example_segment()));
-}
-
-std::string read_in_store(AzureStorage& store, std::string symbol) {
-    auto variant_key = get_variant_key(symbol);
-    auto opts = ReadKeyOpts{};
-    auto result = store.read(std::move(variant_key), opts);
-    return std::get<StringId>(result.atom_key().id());
-}
-
-bool exists_in_store(AzureStorage& store, std::string symbol){
-    auto variant_key = get_variant_key(symbol);
-    return store.key_exists(variant_key);
-}
-
-void remove_in_store(AzureStorage& store, std::vector<std::string> symbols){
-    auto to_remove = std::vector<VariantKey>();
-    for (auto& symbol : symbols){
-        to_remove.emplace_back(get_variant_key(symbol));
-    }
-    auto opts = RemoveOpts();
-    store.remove(Composite(std::move(to_remove)), opts);
-}
-
-std::set<std::string> list_in_store(AzureStorage& store){
-    auto keys = std::set<std::string>();
-    store.iterate_type(KeyType::TABLE_DATA, [&keys](VariantKey&& key){
-        auto atom_key = std::get<AtomKey>(key);
-        keys.emplace(std::get<StringId>(atom_key.id()));
-    });
-    return keys;
-}
 
 TEST_F(AzureMockStorageFixture, test_key_exists) {
     write_in_store(store, "symbol");
